# Application/Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Flux Backend",
    "version": "1.0.0",
    "description": "A comprehensive API management system for environment files, web flows, and file uploads",
    "contact": {
      "name": "API Flux Team",
      "email": "support@apiflux.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.apiflux.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/web-flows/{id}/openapi-docs": {
      "get": {
        "summary": "Get OpenAPI documentation for a web flow",
        "tags": [
          "WebFlows"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The web flow ID to get OpenAPI docs for"
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI documentation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "openApiDocs": {
                      "type": "object",
                      "description": "The OpenAPI specification document"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "server",
                        "file"
                      ],
                      "description": "The source of the OpenAPI documentation"
                    },
                    "sourceDetails": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string",
                          "description": "Server URL (if source is server)"
                        },
                        "fileName": {
                          "type": "string",
                          "description": "File name (if source is file)"
                        },
                        "fileId": {
                          "type": "integer",
                          "description": "File ID (if source is file)"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "OpenAPI configuration not enabled or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OpenAPI configuration is not enabled for this web flow"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Web flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Web flow not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}